Now Allowing Indexing by Google and other Search Engines
All user-agents get Events, but only known browsers
get Messages

This is because search engine crawlers need to see some content,
and Events are already out there and searchable on the web.

Messages are a bit more private, and their ephemeral nature means
it makes more sense for the search engine not to see them.